// src/shared/api/generated/model/updateCreatorRatesRequestChannelRatesItemFormat.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Marz API
 * Backend de Marz. Fuente de verdad del contrato HTTP entre marz-api y marz-front.
Se consume con `oapi-codegen` (server) y `openapi-typescript` + `openapi-fetch` (client).

 * OpenAPI spec version: 0.1.0
 */

export type UpdateCreatorRatesRequestChannelRatesItemFormat = typeof UpdateCreatorRatesRequestChannelRatesItemFormat[keyof typeof UpdateCreatorRatesRequestChannelRatesItemFormat];


export const UpdateCreatorRatesRequestChannelRatesItemFormat = {
  ig_reel: 'ig_reel',
  tiktok_video: 'tiktok_video',
  yt_short: 'yt_short',
} as const;